requirements.txt
streamlit
boto3
tavily-python
python-dotenv
tenacity